requirements.txt
generativeai
streamlit 
python-dotenv
google.generativeai